requirements.txt
fastapi
uvicorn
python-dotenv
groq
sqlalchemy
streamlit
requests
psycopg2-binary
mysql-connector-python